refactor
fix: Исправлены замечания, добавлен yml файл с ссылкой на GitHub тесты.
