+++ .github/workflows/api-tests.yml
@@ -1,0 +1,8 @@
+name: Java Tests
+
+on:
+  pull_request:
+
+jobs:
+  build:
+    uses: yandex-praktikum/java-kanban/.github/workflows/api-tests.yml@ci
